[ng] maint: downgrade version: 2.0a -> 0.5a
Since Automake is still in the version series 1.x (at the moment of
writing the last stable version being 1.11.3), labelling Automake-NG
with a version of "2.0a" (as we've done in a recent commit) would
mud the waters about what the "real" or "more recent" Automake is,
and cause even more confusion about the Automake vs. Automake-NG
distinction.  Which is precisely what we don't want.

So let's downgrade the Automake-NG version to "0.5a".

This change comes with a caveat of its own, though.

After this "version downgrade", a project which requires a minimal
Automake API version in AM_INIT_AUTOMAKE or AUTOMAKE_OPTIONS might
get a "spurious" warning if someone tries to use Automake-NG to
bootstrap it.  For example, a project whose configure.ac contains:

  AM_INIT_AUTOMAKE([1.10])

would trigger a complaint from Automake-NG, since "1.10" is
interpreted as a version number greater than "0.5a".

Then again, while this is annoying right now, in the long run it
might turn out to actually be helpful, in case we introduce some
big or subtle backward-incompatibility between Automake-NG and
mainstream Automake.

Suggestion by Dave Hart.

* configure.ac (AC_INIT): Change version number from 2.0a to 0.5a.