tests: avoid FreeBSD make VPATH issues in more tests (see bug#7884)
See also similar change 'v1.11-755-g818bc40' of 07-02-2012,
"tests: work around bug#7884 in many yacc/lex tests".  The
rationale for this change is basically the same.

* tests/lex-lib-external.test: Use "yl_distcheck" instead of
bare "$MAKE distcheck" to avoid extra failures caused by
automake bug#7884.
* tests/lex-noyywrap.test: Likewise.
* tests/lex-libobj.test: Likewise.
* tests/man6.test: This test suffers from the same FreeBSD make
incompatibility in VPATH handling that is the source of automake
bug#7884.  Since this is caused by rules that are defined in the
Makefile.am by the test itself, rather that being generated by
automake, the best thing to do is to skip this test if we detect
the presence of said make incompatibility (through the function
'useless_vpath_rebuild').
* tests/man8.test: Likewise.