tests: be stricter in determining whether "make -jN" works
* tests/parallel-tests3.test: With at least Solaris 10 CCS make,
"make -jN" does not fail, but merely prints a warning about
"DistributedMake -j option" being ignored.  This was causing this
test case to fail spuriously.  So we now use a stricter check in
determining whether make works in parallel mode.  Since we are at
it, we try to cater also to Sun Distributed make, which understands
"-j N" but *not* "-jN".