aclocal: don't create /usr/share/aclocal as empty
Related to automake bug#10791.

Since automake 1.11.2, aclocal no longer succeeds if the system
acdir '${datadir}/aclocal' (i.e., "/usr/share/aclocal" when
${prefix} is "/usr") does not exist.  So we now have a directory
*installed empty by default*, but whose existence is required by
aclocal in order for it to function at all.  This might be
confusing, and sounds like asking for trouble in the long term.

So we now also install a simple README file in that directory,
which briefly explains its purpose and warn that its existence
is required by the normal aclocal functioning.

* m4/acdir/README: New file.
* m4/Makefile.am (dist_system_ac_DATA): Add it.
* m4/acdir/.placeholder: Delete, no more needed.
* m4/Makefile.am (EXTRA_DIST): Remove it.
* THANKS: Update.