docs: improve "make distcheck" documentation
* doc/automake.texi (Checking the Distribution): The fact that
"make distcheck" uses a read-only srcdir wasn't documented clearly
enough here.  Fix that.  Since we are at it, be more explicit
about the steps undertaken by "make distcheck", and re-organize
this section a bit.  Motivated by the discussion on automake
bug#10878.
(distcleancheck): Rename this node (in the "FAQ" section) ...
(Errors with distclean): ... to this.
Add a proper "FIXME" comment.