self checks: relax tests on cleanup
Some find(1) implementations have problems operating recursively on
directories having subdirectories with null permissions, even when
the permissions of such subdirectory should be fixed by find before
it descends into them; for example, with this setup:

  % mkdir a a/b
  % chmod 000 a/b

a command like this:

  % find a -type d ! -perm -700 -exec chmod u+rwx '{}' ';'

fails with this diagnostic on MacOS X 10.7:

  find: a/b: Permission denied

and with this diagnostic on Solaris 10:

  find: cannot read dir a/b: Permission denied

The problem is that our self checks were simply demanding too much
from our cleanup trap: our tests never use subdirectories with null
permissions, so it doesn't matter if the cleanup trap fails to
handle those.  Just relax the self checks to avoid such useless
testsuite noise.

* tests/self-check-cleanup.tap: Only try directories missing
write permissions, not with null permission.  That should be
enough for our usages.