tests: do not clobber the $CC set by AM_PROG_CC_C_O
If $CC is originally a losing compiler, AM_PROG_CC_C_O will
modify $CC.  "$MAKE -e" will then clobber this modified $CC
and reintroduce the raw losing compiler causing the test
to fail, as subdir-objects is in effect.

* tests/yacc-dist-nobuild-subdir.test: Drop -e from the $MAKE
invocations.  This is safe since DISTCHECK_CONFIGURE_FLAGS is not
specified in the Makefile and since the YACC override is carried
over from the configure invocation into the Makefile making it
redundant to specifiy it again in the environment for $MAKE.