maint: update TODO file
We don't have to make creation of $(foodir) conditional when
foo_PRIMARY is conditionally defined; that has been solved in
a more sweeping way by commit v1.11-759-g368f1c4 of March 18,
which solved bug#11030 and bug#10997.

The "check" target already depends on the "all" target.

Remove comment "investigate problems with conditionally defined
libraries"; it is aimed at automake 1.5 (!) and too terse and old
to be useful now.

Remove the 12 years old suggestion that "distcheck should make sure
that each file that uses _() is listed in POTFILES.in"; this is
handled by gnulib-provided syntax check 'sc_po_check' already today.

Don't tell to "Run automake before libtool"; automake does not
automatically run libtool anymore today.

Remove hint about the possibility to "rewrite automake in guile" or
"add a GUI": we are not going to do so.

The automatic de-ANSI-fication feature has already been remove, so
no need to suggest its potential removal in the future.

Remove reference to a prospective "cvs" option that could add some
cvs-specific rules.  First of all, CVS is falling out of use today.
Secondly, and most to the point, instead of guessing which VCS
is going to be most popular and try to hack support for it into
automake, we should leave it to developer-specific makefile
fragments offered by projects like Gnulib to provide such rules
(they are already doing so with succcess).

About DOS support: it should actually be dropped, not extended.

Add link about Karl Berry's proposal of preferring HTML over Info
documentation installed in the local system:
<http://lists.gnu.org/archive/html/texinfo-devel/2012-03/msg00018.html>

Remove items about "beautification" of output Makefile.in; they
are pretty obvious, and mostly noise (after all, everybody prefers
clean output where possible).  Leave it to any developer doing
future refactorings to find out low-hanging and/or high-impact
optimizations and beatifications of the generated Makefiles.

Remove item about internationalization: we're not going to
implement it.

Remove an item explicitly marked as "not that it matters".

The order of files in a distribution tarball does not really
matter, nobody has ever complained about it, so do not worry
about it.

Remove the suggestions about allowing hierarchy of dirs to share
one aclocal.m4 and about printing full file name of Makefile.am
or configure.in when giving error, in order to help very large
trees: nobody has ever complained about the current behaviours,
so no real need to change them.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>