[ng] tests: prefer to grep make stderr for expected diagnostic
Few make implementations, like BSD's, can print diagnostic from make or
its spawned recipes on stdout rather than on stderr.  Some of our tests
had been relaxed to cater for this.  Now that we assume GNU make, there
is no reason to be this "sloppy" anymore.

* t/uninstall-fail.sh: Expect diagnostic from make and its recipes
to be on stderr, not possibly also on stdout.
* t/lisp3.sh: Likewise.
* t/test-missing.sh: Likewise.
* t/test-missing2.sh: Likewise.
* t/distcheck-pr10470.sh: Likewise.
* t/distcheck-pr9579.sh: Likewise.
* t/dist-missing-am.sh: Likewise.
* t/dist-missing-included-m4.sh: Likewise.
* t/dist-missing-m4.sh: Likewise.
* t/deleted-am.sh: Likewise.
* t/deleted-m4.sh: Likewise.
* t/distcheck-configure-flags-am.sh: Likewise.
* t/distcheck-configure-flags-subpkg.sh: Likewise.
* t/distcheck-configure-flags.sh: Likewise.
* t/yacc-dist-nobuild.sh: Likewise.  Also, be stricter in matching
expected make diagnostic.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>