[ng] parallel-tests: remove workaround for BSD make
When BSD make is run in parallel mode, it apparently strips any
leading directory component from the automatic variable '$*' (of
course, against what POSIX mandates).  In order to avoid this
behaviour from breaking the parallel-tests driver, we added an
hacky workaround in commit v1.11-1982-ge1606b3 of 18-02-2012.
Now that we assume GNU make, this workaround is not needed anymore,
so let's remove it.

* lib/check2.am (am__set_b): Remove its definition and its uses.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>