automake: refactor pre-processing of makefile fragments
This change will provide the automake script with a new function that
reads in a Makefile fragment *without* performing Automake ad-hoc parsing,
but only the pre-processing step, i.e., removal of '##' comments and
substitution of tokens like '%SUBDIRS%', '%?LIBTOOL%' or '?GENENRIC?'.

This will very likely be useful for the work on Automake-NG.

This is a pure refactoring, with no intended functional or semantic
changes.

* automake.in (preprocess_file): New function, extracted ...
(make_paragraphs): ... from here.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>