[ng] silent rules: support for them is always active now
At the time of writing, in mainstream Automake, support for silent rules
is optional and activated only if the 'silent-rules' option is specified.

The rationale behind this behaviour is that the silent-rules machinery
originally [1] only worked with make implementations supporting nested
variables expansions, which wasn't a POSIX-specified features, so that
packages wanting to be "extra-portable" couldn't use the silent-rules
machinery at all.

But since Automake-NG only targets GNU make, and since silent rules are
disabled by default even when the 'silent-rules' option is given [2], we
can now have the support for silent rules always enabled.

The 'silent-rules' option will thus become a no-op, but will reaming a
valid option for the time being, for better backward-compatibility.

[NOTES]

  [1] At least, until Paul Eggert's commit 'v1.11-598-g8493499' of
      25-12-2011, "silent-rules: fallback for makes without nested
      vars", motivate by automake bugs #9928 and #10237.

  [2] Unless the developer makes them enabled by default be calling
      "AM_SILENT_RULES([yes])" in configure.ac.

[DETAILS]

For more details, see this thread on the automake-ng mailing list:
<http://lists.gnu.org/archive/html/automake-ng/2012-04/msg00027.html>

* automake.in: Enable silent-rules machinery unconditionally.  Reword
some comments.
* m4/init.m4: Call 'AC_REQUIRE' on 'AM_SILENT_RULES' unconditionally.
* lib/Automake/Options.pm: Recognize 'silent-rules' as a dummy option
* t/silent.sh: Remove now-redundant calls to AM_SILENT_RULES.
* t/silent2.sh: Likewise.
* t/silent3.sh: Likewise.
* t/silent4.sh: Likewise.
* t/silent6.sh: Likewise.
* t/silent8.sh: Likewise.
* t/silent9.sh: Likewise.
* t/silent-lex.sh: Likewise.
* t/silent-many-gcc.sh: Likewise.
* t/silent-many-generic.sh: Likewise.
* t/silent-yacc-headers.sh: Likewise.
* t/silent-yacc.sh: Likewise.
* t/silentcxx-gcc.sh: Likewise.
* t/silentcxx.sh: Likewise.
* t/silentf77.sh: Likewise.
* t/silentf90.sh: Likewise.
* t/help-silent.sh: Likewise.
* t/silent-configsite.sh: Likewise, and other minor adjustments.
* t/silent7.sh: Likewise.  Also, extend to check that silent rules
are disabled by default, and remove obsoleted checks about the
'silent-rules' option being rejected in AUTOMAKE_OPTIONS.
* t/compile_f_c_cxx.sh: Adjust to avoid a spurious failure.
* t/silent-amopts.sh: Remove as obsolete.
* t/list-of-tests.mk: Adjust.
* NG-NEWS: Update.
* doc/automake.texi: Update and simplify accordingly.
* configure.ac (AM_INIT_AUTOMAKE): Remove 'silent-rules' option.

Signed-off-by: Stefano Lattarini <stefano.lattarini@gmail.com>